Moved graphic elements' js to separate file
