+++ index.html
@@ -19,10 +19,16 @@
 					<!--  <h1>Debug menu</h1> -->
 					<!-- <div id="counter" class="debugOpt">0</div> -->
 					<button id="audioButton" class="debugOpt" type="button">Toggle audio</button>
+					<!--  <button id="pauseButton" class="debugOpt" type="button">Pause</button> -->
 				</div>
 			</div>
 			<div id="startMenu">
 				<h3 class="startOpt">Press play to start!</h3>
+				<select class="startOpt" name="select">
+					<option value="value1" selected>Value 1</option>
+					<option value="value2">Value 2</option>
+					<option value="value3">Value 3</option>
+				</select>
 				<button id="playButton" class="startOpt" type="button">Play!</button>
 			</div>
 		</div>
